requirements.txt
# requirements.txt for DefectDojo using Python 3.x
asteval==0.9.29
bleach==6.0.0
bleach[css]
celery==5.2.7
coverage==7.2.4
defusedxml==0.7.1
django_celery_results==2.5.0
django-auditlog==2.2.2
django-dbbackup==4.0.2
django-environ==0.10.0
django-filter==23.1
django-imagekit==4.1.0
# This library is very outdated, but is a pillar of DefectDojo
# django-multiselectfield==0.1.12
git+https://github.com/DefectDojo/django-multiselectfield@master#egg=django-multiselectfield
django-polymorphic==3.1.0
django-crispy-forms==2.0
django_extensions==3.2.1
django-slack==5.19.0
# This library is very outdated and not directly. It is used solely for migration
# purposes to django-tagulous, so it must stay
# django-tagging==0.5.0
git+https://github.com/DefectDojo/django-tagging@develop#egg=django-tagging
django-watson==1.6.3
django-prometheus==2.2.0
Django==4.1.9
djangorestframework==3.14.0
gunicorn==20.1.0
html2text==2020.1.16
humanize==4.6.0
jira==3.5.0
PyGithub==1.58.1
lxml==4.9.2
Markdown==3.4.3
mysqlclient==2.1.1
openpyxl==3.1.2
xlrd==1.2.0
Pillow==9.5.0  # required by django-imagekit
psycopg2-binary==2.9.6
cryptography==40.0.2
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.29.0
sqlalchemy==1.4.46  # Required by Celery broker transport
supervisor==4.2.5
urllib3==1.26.11
uWSGI==2.0.21
vobject==0.9.6.1
whitenoise==5.2.0
titlecase==2.3
social-auth-app-django==5.2.0
social-auth-core==4.4.2
Python-jose==3.3.0
gitpython==3.1.31
debugpy==1.6.7
python-gitlab==3.14.0
drf_yasg==1.21.5
cpe==1.2.1
packageurl-python==0.11.1
django-crum==0.7.9
JSON-log-formatter==0.5.2
django-split-settings==1.2.0
django-debug-toolbar==4.0.0
django-debug-toolbar-request-history==0.1.4
vcrpy==4.2.1
vcrpy-unittest==0.1.7
django-tagulous==1.3.3
PyJWT==2.6.0
cvss==2.6
django-fieldsignals==0.7.0
hyperlink==21.0.0
django-test-migrations==1.2.0
djangosaml2==1.5.6
drf-spectacular==0.26.2
django-ratelimit==4.0.0
argon2-cffi==21.3.0
blackduck==1.1.0
pycurl==7.45.2  # Required for Celery Broker AWS (SQS) support
boto3==1.26.122  # Required for Celery Broker AWS (SQS) support
netaddr==0.8.0
vulners==2.0.9
fontawesomefree==6.4.0
docxtpl==0.16.7
beautifulsoup4==4.12.2
